Add water flowing and organize
- Add WATER_SOURCE block and change all world generation functions to use WATER_SOURCE instead of B_WATER7
- Add water flowing in world.update()
- Add world.update5Tick to match the tickrate of flowing water in Minecraft
- Add floor and ceil functions to find the smaller and larger of a pair of numbers, respectively
- Move update functions and generate functions into worldUpdate.cpp and worldGenerate.cpp, respectively
- Rename comTemplates.h to comT.h
- Add miscT.h
